requirements.txt
streamlit
pandas
python-barcode
openpyxl
